Refactoring `LocaleSelect` & `SystemStore`.
